Module tendermint_rpc::abci

Old ABCI structures, formerly defined in tendermint::abci.

The original contents of tendermint::abci were created only to model RPC responses, not to model ABCI itself:

NOTE: This module contains types for ABCI responses as consumed from RPC endpoints. It does not contain an ABCI protocol implementation.

The old types should be eliminated and merged with the new ABCI domain types. Moving them here in the meantime disentangles improving the ABCI domain modeling from changes to the RPC interface.

Re-exports

pub use self::responses::DeliverTx;

pub use self::responses::Event;

pub use self::responses::Responses;

pub use self::transaction::Transaction;

Modules

responses

ABCI response types used by the /block_results RPC endpoint.

tag

Tags

transaction

Transactions

Structs

Data

ABCI transaction data.

Gas

Gas: representation of transaction processing resource costs

Info

ABCI info

Log

ABCI log data

Path

Path to ABCI data

Enums

Code

ABCI application response codes.