tendermint_rpc::client::http
Struct HttpClient

pub struct HttpClient { /* private fields */ }

A JSON-RPC/HTTP Tendermint RPC client (implements crate::Client).

Supports both HTTP and HTTPS connections to Tendermint RPC endpoints, and allows for the use of HTTP proxies (see HttpClient::new_with_proxy for details).

Does not provide crate::event::Event subscription facilities (see crate::WebSocketClient for a client that does).

Examples

use tendermint_rpc::{HttpClient, Client};

#[tokio::main]
async fn main() {
    let client = HttpClient::new("http://127.0.0.1:26657")
        .unwrap();

    let abci_info = client.abci_info()
        .await
        .unwrap();

    println!("Got ABCI info: {:?}", abci_info);
}

Implementations

impl HttpClient

pub fn new_from_parts(inner: Client, url: Url, compat: CompatMode) -> Self

Construct a new Tendermint RPC HTTP/S client connecting to the given URL. This avoids using the Builder and thus does not perform any validation of the configuration.

pub fn new<U>(url: U) -> Result<Self, Error>
where U: TryInto<HttpClientUrl, Error = Error>,

Construct a new Tendermint RPC HTTP/S client connecting to the given URL.

pub fn new_with_proxy<U, P>(url: U, proxy_url: P) -> Result<Self, Error>
where U: TryInto<HttpClientUrl, Error = Error>, P: TryInto<HttpClientUrl, Error = Error>,

Construct a new Tendermint RPC HTTP/S client connecting to the given URL, but via the specified proxy’s URL.

If the RPC endpoint is secured (HTTPS), the proxy will automatically attempt to connect using the HTTP CONNECT method.

pub fn builder(url: HttpClientUrl) -> Builder

Initiate a builder for a Tendermint RPC HTTP/S client connecting to the given URL, so that more configuration options can be specified with the builder.

pub fn set_compat_mode(&mut self, compat: CompatMode)

Set compatibility mode on the instantiated client.

As the HTTP client is stateless and does not support subscriptions, the protocol version it uses can be changed at will, for example, as a result of version discovery over the /status endpoint.