Module tendermint_rpc::endpoint::broadcast::tx_commit

/broadcast_tx_commit: only returns error if mempool.CheckTx() errs or if we timeout waiting for tx to commit.

Modules

v0_34

Serialization for /broadcast_tx_commit endpoint format in Tendermint 0.34

v0_37

Serialization for /broadcast_tx_commit endpoint format in CometBFT 0.37

Structs

Request

/broadcast_tx_commit: only returns error if mempool.CheckTx() errs or if we timeout waiting for tx to commit.

Response

Response from /broadcast_tx_commit.