Module tendermint_rpc::endpoint::consensus_state

/consensus_state endpoint JSON-RPC wrapper

Structs

Fingerprint

HeightRoundStep

A compound object indicating a height, round and step for consensus state.

Request

Get the current consensus state.

Response

The current consensus state (UNSTABLE).

RoundState

The state of a particular consensus round.

RoundVotes

Details of all votes for a particular consensus round.

ValidatorInfo

VoteSummary

Enums

RoundVote

Details of a single vote from a particular consensus round.