Module tendermint_rpc::query

Structured querying for the Tendermint RPC event subscription system.

See Query for details as to how to construct queries.

Structs

Query

A structured query for use in interacting with the Tendermint RPC event subscription system.

Enums

Condition

The different types of conditions supported by a Query.

EventType

The types of Tendermint events for which we can query at present.

Operand

A typed operand for use in an Condition.

Term

A term in a query is either an event type or a general condition. Exclusively used for query parsing.