Struct tendermint_rpc::query::Query

pub struct Query { /* private fields */ }

A structured query for use in interacting with the Tendermint RPC event subscription system.

Allows for compile-time validation of queries.

See the subscribe endpoint documentation for more details.

Examples

Direct construction of queries

use tendermint_rpc::query::{Query, EventType};

let query = Query::from(EventType::NewBlock);
assert_eq!("tm.event = 'NewBlock'", query.to_string());

let query = Query::from(EventType::Tx).and_eq("tx.hash", "XYZ");
assert_eq!("tm.event = 'Tx' AND tx.hash = 'XYZ'", query.to_string());

let query = Query::from(EventType::Tx).and_gte("tx.height", 100_u64);
assert_eq!("tm.event = 'Tx' AND tx.height >= 100", query.to_string());

Query parsing

use tendermint_rpc::query::{Query, EventType};

let query: Query = "tm.event = 'NewBlock'".parse().unwrap();
assert_eq!(query, Query::from(EventType::NewBlock));

let query: Query = "tm.event = 'Tx' AND tx.hash = 'XYZ'".parse().unwrap();
assert_eq!(query, Query::from(EventType::Tx).and_eq("tx.hash", "XYZ"));

let query: Query = "tm.event = 'Tx' AND tx.height >= 100".parse().unwrap();
assert_eq!(query, Query::from(EventType::Tx).and_gte("tx.height", 100_u64));

Implementations

impl Query

pub fn eq(key: impl ToString, value: impl Into<Operand>) -> Self

Query constructor testing whether <key> = <value>

pub fn lt(key: impl ToString, value: impl Into<Operand>) -> Self

Query constructor testing whether <key> < <value>

pub fn lte(key: impl ToString, value: impl Into<Operand>) -> Self

Query constructor testing whether <key> <= <value>

pub fn gt(key: impl ToString, value: impl Into<Operand>) -> Self

Query constructor testing whether <key> > <value>

pub fn gte(key: impl ToString, value: impl Into<Operand>) -> Self

Query constructor testing whether <key> >= <value>

pub fn contains(key: impl ToString, value: impl ToString) -> Self

Query constructor testing whether <key> CONTAINS <value> (assuming key contains a string, this tests whether value is a sub-string within it).

pub fn exists(key: impl ToString) -> Self

Query constructor testing whether <key> EXISTS.

pub fn and_eq(self, key: impl ToString, value: impl Into<Operand>) -> Self

Add the condition <key> = <value> to the query.

pub fn and_lt(self, key: impl ToString, value: impl Into<Operand>) -> Self

Add the condition <key> < <value> to the query.

pub fn and_lte(self, key: impl ToString, value: impl Into<Operand>) -> Self

Add the condition <key> <= <value> to the query.

pub fn and_gt(self, key: impl ToString, value: impl Into<Operand>) -> Self

Add the condition <key> > <value> to the query.

pub fn and_gte(self, key: impl ToString, value: impl Into<Operand>) -> Self

Add the condition <key> >= <value> to the query.

pub fn and_contains(self, key: impl ToString, value: impl ToString) -> Self

Add the condition <key> CONTAINS <value> to the query.

pub fn and_exists(self, key: impl ToString) -> Self

Add the condition <key> EXISTS to the query.