Trait tendermint_rpc::MockRequestMatcher

pub trait MockRequestMatcher: Send + Sync {
    fn response_for<R>(&self, request: R) -> Option<Result<R::Response, Error>>
    where
        R: Request;
}

A trait required by the MockClient that allows for different approaches to mocking responses for specific requests.

Required methods

fn response_for<R>(&self, request: R) -> Option<Result<R::Response, Error>> where
R: Request,

Provide the corresponding response for the given request (if any).