Crate tendermint_rpc

Tendermint RPC definitions and types.

Client

This crate optionally provides access to different types of RPC client functionality and different client transports based on which features you select when using it.

Several client-related features are provided at present:

http-client - Provides HttpClient, which is a basic RPC client that interacts with remote Tendermint nodes via JSON-RPC over HTTP or HTTPS. This client does not provide event::Event subscription functionality. See the Tendermint RPC for more details.

websocket-client - Provides WebSocketClient, which provides full client functionality, including general RPC functionality as well as event::Event subscription functionality. Can be used over secure (wss://) and unsecure (ws://) connections.

Mock Clients

Mock clients are included when either of the http-client or websocket-client features are enabled to aid in testing. This includes MockClient, which implements both Client and SubscriptionClient traits.

Re-exports

pub use client::Client;

pub use client::MockClient;

pub use client::MockRequestMatcher;

pub use client::MockRequestMethodMatcher;

pub use client::Subscription;

pub use client::SubscriptionClient;

pub use client::HttpClient;

pub use client::HttpClientUrl;

pub use client::WebSocketClient;

pub use client::WebSocketClientDriver;

pub use client::WebSocketClientUrl;

pub use error::Error;

pub use request::Request;

pub use request::SimpleRequest;

pub use response::Response;

pub use response_error::Code;

pub use response_error::ResponseError;

Modules

client

Tendermint RPC client.

dialect

Helper types to generalize differences in serialization between Tendermint RPC protocol versions.

endpoint

Tendermint JSON-RPC endpoints

error

JSON-RPC error types

event

RPC subscription event-related data structures.

query

Structured querying for the Tendermint RPC event subscription system.

request

JSON-RPC requests

response

JSON-RPC response types

response_error

serializers

Serde serializers

Structs

PageNumber

A page number in paginated RPC responses.

PerPage

The number of items to return per page, for paginated RPC responses.

Url

A uniform resource locator (URL), with support for only those schemes/protocols supported by Tendermint RPC clients.

Version

JSON-RPC version

WebSocketConfig

The configuration for WebSocket connection.

Enums

Id

JSON-RPC ID: request-specific identifier

Method

JSON-RPC request methods.

Order

Ordering of paginated RPC responses.

Paging

Pagination control for those RPC client methods supporting pagination.

Scheme

The various schemes supported by Tendermint RPC clients.