tendermint_rpc
Module client

Tendermint RPC client.

Modules

http

HTTP-based transport for Tendermint RPC Client.

sync

Synchronization primitives specific to the Tendermint RPC client.

websocket

WebSocket-based clients for accessing Tendermint RPC functionality.

Structs

HttpClient

A JSON-RPC/HTTP Tendermint RPC client (implements crate::Client).

HttpClientUrl

A URL limited to use with HTTP clients.

Subscription

An interface that can be used to asynchronously receive Events for a particular subscription.

WebSocketClient

Tendermint RPC client that provides access to all RPC functionality (including Event subscription) over a WebSocket connection.

WebSocketClientDriver

Drives the WebSocket connection for a WebSocketClient instance.

WebSocketClientUrl

A URL limited to use with WebSocket clients.

WebSocketConfig

Low-level WebSocket configuration The configuration for WebSocket connection.

Enums

CompatMode

Protocol compatibility mode for a Tendermint RPC client.

Traits

Client

Provides lightweight access to the Tendermint RPC. It gives access to all endpoints with the exception of the event subscription-related ones.

SubscriptionClient

A client that exclusively provides Event subscription capabilities, without any other RPC method support.