Module tendermint_rpc::client

Tendermint RPC client.

Modules

sync

Synchronization primitives specific to the Tendermint RPC client.

Structs

HttpClient

A JSON-RPC/HTTP Tendermint RPC client (implements crate::Client).

HttpClientUrl

A URL limited to use with HTTP clients.

MockClient

A mock client implementation for use in testing.

MockRequestMethodMatcher

Provides a simple MockRequestMatcher implementation that simply maps requests with specific methods to responses.

Subscription

An interface that can be used to asynchronously receive Events for a particular subscription.

WebSocketClient

Tendermint RPC client that provides access to all RPC functionality (including Event subscription) over a WebSocket connection.

WebSocketClientDriver

Drives the WebSocket connection for a WebSocketClient instance.

WebSocketClientUrl

A URL limited to use with WebSocket clients.

WebSocketConfig

The configuration for WebSocket connection.

Enums

CompatMode

Protocol compatibility mode for a Tendermint RPC client.

Traits

Client

Provides lightweight access to the Tendermint RPC. It gives access to all endpoints with the exception of the event subscription-related ones.

MockRequestMatcher

A trait required by the MockClient that allows for different approaches to mocking responses for specific requests.

SubscriptionClient

A client that exclusively provides Event subscription capabilities, without any other RPC method support.