Trait tendermint_rpc::request::SimpleRequest

pub trait SimpleRequest: Request { }

Simple JSON-RPC requests which correlate with a single response from the remote endpoint.

An example of a request which is not simple would be the event subscription request, which, on success, returns a Subscription and not just a simple, singular response.