Trait tendermint_rpc::request::SimpleRequest

pub trait SimpleRequest<S: Dialect = LatestDialect>: Request<S> {
    type Output: From<Self::Response>;
}

Simple JSON-RPC requests which correlate with a single response from the remote endpoint.

An example of a request which is not simple would be the event subscription request, which, on success, returns a Subscription and not just a simple, singular response.

Required Associated Types

type Output: From<Self::Response>

The output data, converted from Response.

Object Safety

This trait is not object safe.