Module tendermint_rpc::serializers

Serde serializers

Serializers and deserializers for a transparent developer experience.

CAUTION: There are no guarantees for backwards compatibility, this module should be considered an internal implementation detail which can vanish without further warning. Use at your own risk.

Modules

bytes

Serialize/deserialize bytes (Vec<u8>) type

evidence

from_str

Serialize and deserialize any T that implements [core::str::FromStr] and [core::fmt::Display] from or into string. Note this can be used for all primitive data types.

nullable

Serialize/deserialize nilable type into T, where nil turns into the default impl.

opt_tm_hash_base64

Encoding/decoding Option Tendermint hashes to/from base64.

optional

Serialize/deserialize Option<T> type where T has a serializer/deserializer. Use null if the received value equals the Default implementation.

optional_from_str

De/serialize an optional type that must be converted from/to a string.

part_set_header_total

Serialize and deserialize part_set_header.total (from string or u32), (into u32 in part_set_header.total).

time_duration

Serialize/deserialize core::time::Duration type from and into string:

timestamp

Serialize/deserialize Timestamp type from and into string:

tm_hash_base64

Encoding/decoding Tendermint hashes to/from base64.

tx_hash_base64

Encoding/decoding ABCI transaction hashes to/from base64.

txs

Serialize/deserialize Vec<Vec<u8>> type from and into transactions (Base64String array).